fix: show deep files in experiment code viewer (#7945)
This fixes an issue where files more than one folder deep were not
viewable in the experiment code view. The code to find which file node
in the file list is active assumed that the node keys would only contain
the last path segment (i.e: a file at `foo/bar/baz` would have a key of
`baz`) or fully match the selected file path. For files 1 folder deep,
the search would match the folder, then fully match the file, but
because deeper files needed to match more of the partial path, it would fail.

This is fixed by instead matching the current path with all previously
matched segments with the node key. Because the only consumer of
CodeEditor that this effects is ExperimentCodeViewer, this should be fine.