fix: don't reinstall playwright when generating screenshots
this fixes a subtle error that happens in the screenshot differ. when
generating upstream screenshots we install fresh copies of `playwright`
and `playwright-core`. this causes issues when a new version of
playwright is released, because the upstream then can't find the
installed browsers. We did this because the upstream initially didn't
have playwright installed -- now that it does, we instead use the cached
dependencies from that version of the build. this requires splitting up
the upstream portion of the screenshot generator.