fix: default designkit theme to system theme
this fixes an issue where the design kit diff checker didn't take dark
mode screenshots. when implementing the theme context, we defaulted the
design kit to light mode. this broke the diff checker, because it takes
dark mode screenshots by changing the colorscheme preference of the
browser instead of interacting with the page. We fix this by defaulting
to system mode instead.