fix: ensure card's children stretch to grow
this fixes an issue where the children of the card component did not
take up all of the available space in the card. this is because the flex
behavior of the card was broken when we introduced the surface
component. to fix, we add a new flex container for the chidren and card actions.