chore: add elevation to LineChart and ChartGrid (#72)
* chore: add elevation to LineChart and ChartGrid

this adds elevation to the LineChart component. the linechart container
is now a surface, ensuring it appears at one level above the current
elevation. Because linecharts have borders now, this revealed a bug
where the legend of a chart would appear outside of the chart's borders.
To fix, we now dynamically assign the height of the row based on the
series in the row.

* clean up haspopulatedseries

* john feedback 1